Remove data module, move content to ureg/ and isotopes/
This gives a generally cleaner structure: data files live in /data/
subdirectories of their relevant packages.

Annoyingly it does bring back some of the import-loop headache for
legacy data loader; isotopes/_legacy.py can't be loaded independently
of isotopes/__init__.py which creates a circular dependence on
euphonic.util.  The long-term fix is clear (drop isotopes._legacy) and
the real sin was having functionality of util depend on another module.
